+++ src/main/java/com/chatgpt/studentregistration/repository/UnitRepository.java
@@ -2,5 +2,5 @@ package com.chatgpt.studentregistration.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.chatgpt.studentregistration.model.Unit;
-public interface UnitRepository extends JpaRepository<Unit, Long> {
+public interface UnitRepository extends JpaRepository<Unit, Integer> {
 }
